# docs/openapi.yaml
---
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: User Card Management
    description: Endpoints for users to manage their own cards
  - name: User Profile
    description: Endpoints for managing user profile
paths:
  "/api/user/profile":
    get:
      tags:
        - User Profile
      operationId: getProfile
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponseDTO"
    put:
      tags:
        - User Profile
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserUpdateDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponseDTO"
  "/api/user/profile/password":
    put:
      tags:
        - User Profile
      operationId: updatePassword
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ChangePasswordRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/api/admin/users/{userId}":
    get:
      tags:
        - user-manage-controller
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponseDTO"
    put:
      tags:
        - user-manage-controller
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserUpdateDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponseDTO"
    delete:
      tags:
        - user-manage-controller
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: object
                additionalProperties:
                  type: string
  "/api/user/cards/transfer":
    post:
      tags:
        - User Card Management
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TransferRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/api/auth/register":
    post:
      tags:
        - auth-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RegisterRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/api/auth/login":
    post:
      tags:
        - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LoginRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/AuthResponse"
  "/api/admin/cards/create":
    post:
      tags:
        - card-manage-controller
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CardCreateRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardResponseDTO"
  "/api/user/cards/{id}/block-requested":
    patch:
      tags:
        - User Card Management
      operationId: requestBlock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/api/admin/users/{userId}/block":
    patch:
      tags:
        - user-manage-controller
      operationId: blockUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: object
                additionalProperties:
                  type: string
  "/api/admin/users/{userId}/activate":
    patch:
      tags:
        - user-manage-controller
      operationId: activateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: object
                additionalProperties:
                  type: string
  "/api/admin/cards/{id}/block":
    patch:
      tags:
        - card-manage-controller
      operationId: blockCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/api/admin/cards/{id}/activate":
    patch:
      tags:
        - card-manage-controller
      operationId: activateCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/api/user/cards/{id}":
    get:
      tags:
        - User Card Management
      operationId: getMyCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardResponseDTO"
  "/api/user/cards/{id}/transactions":
    get:
      tags:
        - User Card Management
      operationId: getCardTransactions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/TransactionResponseDTO"
  "/api/user/cards/{id}/balance":
    get:
      tags:
        - User Card Management
      operationId: getBalance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: number
  "/api/user/cards/transactions":
    get:
      tags:
        - User Card Management
      operationId: getTransactions
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/TransactionResponseDTO"
  "/api/user/cards/transactions/{id}":
    get:
      tags:
        - User Card Management
      operationId: getTransactionDetails
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TransactionResponseDTO"
  "/api/admin/users/{userId}/audit":
    get:
      tags:
        - user-manage-controller
      operationId: getUserAuditHistory
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/AuditLogResponseDTO"
  "/api/admin/users/{id}/cards":
    get:
      tags:
        - user-manage-controller
      operationId: getUserCards
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/CardResponseDTO"
  "/api/admin/users/search":
    get:
      tags:
        - user-manage-controller
      operationId: searchUsers
      parameters:
        - name: filter
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/UserSearchFilter"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PageUserResponseDTO"
  "/api/admin/cards/{id}/history":
    get:
      tags:
        - card-manage-controller
      operationId: getCardHistory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/AuditLogResponseDTO"
  "/api/admin/cards/top-balance":
    get:
      tags:
        - card-manage-controller
      operationId: getTopCards
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/CardResponseDTO"
  "/api/admin/cards/search":
    get:
      tags:
        - card-manage-controller
      operationId: searchCards
      parameters:
        - name: filter
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/CardSearchFilter"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PageCardResponseDTO"
components:
  schemas:
    UserUpdateDTO:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum:
            - ROLE_USER
            - ROLE_ADMIN
    UserResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum:
            - ROLE_USER
            - ROLE_ADMIN
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ChangePasswordRequest:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
          maxLength: 2147483647
          minLength: 6
        confirmPassword:
          type: string
      required:
        - confirmPassword
        - newPassword
        - oldPassword
    TransferRequest:
      type: object
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
      required:
        - firstName
        - lastName
        - password
        - username
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - password
        - username
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
        role:
          type: string
    CardCreateRequestDTO:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        cardType:
          type: string
        category:
          type: string
        bankName:
          type: string
      required:
        - bankName
        - cardType
        - category
        - userId
    CardResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardNumber:
          type: string
        expiryDate:
          type: string
          format: date
        balance:
          type: number
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
            - BLOCK_REQUESTED
        cardType:
          type: string
          enum:
            - STANDARD
            - PLATINUM
            - GOLD
            - VIRTUAL
        cardCategory:
          type: string
          enum:
            - VISA
            - MASTERCARD
            - UZCARD
            - HUMO
            - MIR
        ownerFullName:
          type: string
        createdAt:
          type: string
          format: date-time
    TransactionResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fromCardMasked:
          type: string
        toCardMasked:
          type: string
        amount:
          type: number
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    AuditLogResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        action:
          type: string
        description:
          type: string
        performedBy:
          type: string
        createdAt:
          type: string
          format: date-time
    UserSearchFilter:
      type: object
      properties:
        query:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum:
            - ROLE_USER
            - ROLE_ADMIN
        isActive:
          type: boolean
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    PageUserResponseDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            "$ref": "#/components/schemas/UserResponseDTO"
        number:
          type: integer
          format: int32
        pageable:
          "$ref": "#/components/schemas/PageableObject"
        sort:
          "$ref": "#/components/schemas/SortObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        unpaged:
          type: boolean
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        sort:
          "$ref": "#/components/schemas/SortObject"
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        unsorted:
          type: boolean
        sorted:
          type: boolean
    CardSearchFilter:
      type: object
      properties:
        cardHolderName:
          type: string
        bankName:
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
            - BLOCK_REQUESTED
        cardType:
          type: string
          enum:
            - STANDARD
            - PLATINUM
            - GOLD
            - VIRTUAL
        cardCategory:
          type: string
          enum:
            - VISA
            - MASTERCARD
            - UZCARD
            - HUMO
            - MIR
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    PageCardResponseDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            "$ref": "#/components/schemas/CardResponseDTO"
        number:
          type: integer
          format: int32
        pageable:
          "$ref": "#/components/schemas/PageableObject"
        sort:
          "$ref": "#/components/schemas/SortObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean